tests/wpt/web-platform-tests/css/css-typed-om/the-stylepropertymap/inline/delete-invalid.html

<!doctype html>
<meta charset="utf-8">
<title>Inline StylePropertyMap.delete Error Handling</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#delete-a-stylepropertymap">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../resources/testhelper.js"></script>
<body>
<div id="log">
<script>
'use strict';

test(t => {
  let styleMap = createInlineStyleMap(t, '');
  assert_throws(new TypeError(), () => styleMap.delete('lemon'));
}, 'Deleting an unsupported property name throws a TypeError');

</script>
